requirements.txt
atlassian-python-api @ git+https://github.com/atlassian-api/atlassian-python-api@efcfb44cbb5890ba1f05cad3e52bc3f3e0b8b021
beautifulsoup4==4.14.3
certifi==2025.6.15
charset-normalizer==3.4.2
Deprecated==1.3.1
idna==3.10
jmespath==1.0.1
oauthlib==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.8.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==2.0.1
